% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion.R
\name{criterion}
\alias{criterion}
\title{Tuning parameter selection information criteria}
\usage{
criterion(hatOmega, S, n, crit, ebic.tuning = 0.5)
}
\arguments{
\item{hatOmega}{The estimated precision matrix.}

\item{S}{The sample covariance matrix.}

\item{n}{An integer specifying the sample size.}

\item{crit}{A string specifying the tuning parameter selection criterion to use.
Available options include: \enumerate{
\item "AIC": Akaike information criterion \insertCite{akaike1973information}{fcstat}.
\item "BIC": Bayesian information criterion \insertCite{schwarz1978estimating}{fcstat}.
\item "EBIC": extended Bayesian information criterion \insertCite{foygel2010extended}{fcstat}.
\item "HBIC": high dimensional Bayesian information criterion \insertCite{wang2013calibrating,fan2017high}{fcstat}.
}}

\item{ebic.tuning}{A numeric value (default = 0.5) specifying the tuning parameter to
calculate when \code{crit = "EBIC"}.}
}
\value{
A numeric value.
}
\description{
Provide information criteria for selecting tuning parameters.
}
\references{
\insertAllCited{}
}
