% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deriv.R
\name{deriv}
\alias{deriv}
\title{Derivative of the non-convex regularization penalty}
\usage{
deriv(penalty, Omega, lambda, gamma)
}
\arguments{
\item{penalty}{A string specifying the non-convex penalty to use. Options include:
\enumerate{
\item "adapt": adaptive lasso \insertCite{zou2006adaptive,fan2009network}{fcstat}.
\item "atan": arctangent type penalty \insertCite{wang2016variable}{fcstat}.
\item "exp": exponential type penalty \insertCite{wang2018variable}{fcstat}.
\item "mcp": minimax concave penalty \insertCite{zou2006adaptive}{fcstat}.
\item "scad": smoothly clipped absolute deviation \insertCite{fan2001variable,fan2009network}{fcstat}.
}}

\item{Omega}{The precision matrix.}

\item{lambda}{A numeric value representing the regularization parameter.}

\item{gamma}{A numeric value representing the hyperparameter for the penalty function.
The defaults are: \enumerate{
\item "adapt": 0.5
\item "atan": 0.005
\item "exp": 0.01
\item "mcp": 3
\item "scad": 3.7
}}
}
\value{
A numeric matrix.
}
\description{
Compute the derivative of the specified non-convex regularization penalty.
}
\references{
\insertAllCited{}
}
