% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcstat.est.R
\name{fcstat.est}
\alias{fcstat.est}
\title{Estimation for precision matrix}
\usage{
fcstat.est(
  X,
  method,
  base = "cov",
  lambda = NULL,
  nlambda = 20,
  lambda.min.ratio = NULL,
  gamma = NULL,
  target = 0,
  initial = "glasso",
  utilopt = "flare",
  cores = 1
)
}
\arguments{
\item{X}{\enumerate{
\item An n-by-p data matrix with sample size n and dimension p.
\item A p-by-p sample covariance/correlation matrix with dimension p.
}}

\item{method}{A character string specifying the statistical method for estimating
precision matrix. Available options include: \enumerate{
\item "glasso": graphical lasso \insertCite{friedman2008sparse}{fcstat}.
\item "ridge": graphical ridge \insertCite{vanwieringen2016ridge}{fcstat}.
\item "elnet": graphical elastic net \insertCite{zou2005regularization}{fcstat}.
\item "clime": constrained L1-minimization for inverse (covariance) matrix estimation
\insertCite{cai2011aconstrained}{fcstat}.
\item "tiger": tuning-insensitive graph estimation and regression
\insertCite{liu2017tiger}{fcstat}, which is only applicable when \code{X} is the n-by-p
data matrix.
\item "adapt": adaptive lasso \insertCite{zou2006adaptive,fan2009network}{fcstat}.
\item "atan": arctangent type penalty \insertCite{wang2016variable}{fcstat}.
\item "exp": exponential type penalty \insertCite{wang2018variable}{fcstat}.
\item "mcp": minimax concave penalty \insertCite{zou2006adaptive}{fcstat}.
\item "scad": smoothly clipped absolute deviation \insertCite{fan2001variable,fan2009network}{fcstat}.
}}

\item{base}{A character string (default = "cov") specifying the calculation base,
either the covariance matrix ("cov") or the correlation matrix ("cor"). This is only
applicable when \code{X} is the n-by-p data matrix.}

\item{lambda}{Grid of non-negative scalars for the regularization parameter.
The default is \code{NULL}, which generates its own \code{lambda} sequence based on
\code{nlambda} and \code{lambda.min.ratio}.}

\item{nlambda}{An integer (default = 20) specifying the number of \code{lambda} values
to be generated when \code{lambda = NULL}.}

\item{lambda.min.ratio}{A scalar specifying the fraction of the maximum \code{lambda}
value \eqn{\lambda_{max}} to generate the minimum \code{lambda} \eqn{\lambda_{min}}.
If \code{lambda = NULL}, the program automatically generates a \code{lambda} grid as a
sequence of length \code{nlambda} in log scale, starting from \eqn{\lambda_{min}} to
\eqn{\lambda_{max}}. The default value is 0.4 for \code{method = "clime"}, 0.1 for
\code{method = "tiger"} and 0.01 for other methods.}

\item{gamma}{Grid of scalars specifying the hyperparameter for the chosen \code{method}.
Default values: \enumerate{
\item "elnet": a sequence from 0.1 to 0.9 with increments of 0.1
\item "adapt": 0.5
\item "atan": 0.005
\item "exp": 0.01
\item "scad": 3.7
\item "mcp": 3
}}

\item{target}{A p-by-p symmetric matrix or a scalar (default = 0) serves as the value
for all elements, specifying the target matrix for \code{method = "ridge"} or
\code{method = "elnet"}.}

\item{initial}{A p-by-p matrix or a p-by-p-by-npara (the number of all combinations of
\code{lambda} and \code{gamma}) array specifying the initial estimate for \code{method}
set to \code{"atan"}, \code{"exp"}, \code{"scad"}, or \code{"mcp"}; or specifying
\eqn{\tilde{\Omega}} of the adaptive weight for \code{method = "adapt"}, calculated as
\eqn{|\tilde{\omega}_{ij}|^{-\gamma}}, where \eqn{\tilde{\Omega} := (\tilde{\omega}_{ij})}.
Some options are also offered when a character string is provided (default "glasso"),
including:
\itemize{
\item "glasso": use the precision matrix estimate derived from the graphical lasso.
\item "invS": use the inverse calculation base matrix if the matrix is invertible.
\item "linshrink": use the precision matrix estimate derived from Ledoit-Wolf linear
shrinakge estimator of the population covariance matrix
\insertCite{ledoit2004well}{fcstat}.
\item "nlshrink": use the precision matrix estimate derived from Ledoit-Wolf non-linear
shrinakge estimator of the population covariance matrix
\insertCite{ledoit2015spectrum,ledoit2017numerical}{fcstat}.
}}

\item{utilopt}{A character string specifying the utility option to use for
\code{method = "clime"}: \itemize{
\item "clime_primaldual": the utility function \code{\link[clime]{clime}} from
the package \code{clime} with the linsolver \code{primaldual}.
\item "clime_simplex": the utility function \code{\link[clime]{clime}} from
the package \code{clime} with the linsolver \code{simplex}.
\item "flare": the utility function \code{\link[flare]{sugm}} from
the package \code{flare}.
}}

\item{cores}{An integer (default = 1) specifying the number of cores to use for
parallel execution.}
}
\value{
An object with S3 class "fcstat.est" containing the following components:
\describe{
\item{hatOmega}{A list of estimated precision matrices for \code{lambda} grid and
\code{gamma} grid.}
\item{lambda}{The actual lambda grid used in the program, corresponding to \code{hatOmega}.}
\item{gamma}{The actual gamma grid used in the program, corresponding to \code{hatOmega}.}
\item{X}{The n-by-p data matrix used in the program.}
\item{S}{The p-by-p calculation base matrix used in the program.}
}
}
\description{
Provide a collection of statistical methods to obtain a series of precision matrix
estimates across the grid of tuning parameters.
}
\note{
For the method \code{tiger}, the estimation process solely relies on the raw n-by-p
data \code{X} and does not utilize the argument \code{base}. This argument is not
applicable for \code{tiger} and will have no effect if provided.
}
\references{
\insertAllCited{}
}
