% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcstat.R
\name{fcstat}
\alias{fcstat}
\title{Statistical methods for estimating precision matrix}
\usage{
fcstat(
  X,
  method,
  base = "cov",
  approach = "smp",
  lambda = NULL,
  nlambda = 50,
  lambda.min.ratio = NULL,
  lambda.min = NULL,
  lambda.max = NULL,
  gamma = NA,
  target = 0,
  initial = "glasso",
  crit = "BIC",
  ebic.tuning = 0.5,
  scaling.ratio = 0.5
)
}
\arguments{
\item{X}{An n-by-p data matrix with sample size n and dimension p.}

\item{method}{A character string specifying the statistical methods for estimating
precision matrix. Available options include: \enumerate{
\item "glasso": graphical lasso \insertCite{friedman2008sparse}{fcstat}.
\item "ridge": graphical ridge \insertCite{vanwieringen2016ridge}{fcstat}.
\item "elnet": graphical elastic net \insertCite{zou2005regularization}{fcstat}.
\item "clime": constrained L1-minimization for inverse (covariance) matrix estimation
\insertCite{cai2011aconstrained}{fcstat}.
\item "tiger": tuning-insensitive graph estimation and regression
\insertCite{liu2017tiger}{fcstat}.
\item "adapt": adaptive lasso \insertCite{zou2006adaptive,fan2009network}{fcstat}.
\item "atan": arctangent type penalty \insertCite{wang2016variable}{fcstat}.
\item "exp": exponential type penalty \insertCite{wang2018variable}{fcstat}.
\item "mcp": minimax concave penalty \insertCite{zou2006adaptive}{fcstat}.
\item "scad": smoothly clipped absolute deviation \insertCite{fan2001variable,fan2009network}{fcstat}.
}}

\item{base}{A character string (default = "cov") specifying the calculation base,
either the covariance matrix ("cov") or the correlation matrix ("cor").}

\item{approach}{A character string (default = "smp") specifying the approach used to
compute the calculation base. Available options include: \enumerate{
\item "smp": sample (conventional approach).
\item "lin": linear shrinkage \insertCite{ledoit2004well}{fcstat}.
\item "nlminb": non-linear shrinkage using the optimization routine "nlminb"
\insertCite{ledoit2015spectrum,ledoit2017numerical}{fcstat}.
\item "nloptr": non-linear shrinkage using the optimization routine "nloptr"
\insertCite{ledoit2015spectrum,ledoit2017numerical}{fcstat}.
}}

\item{lambda}{Grid of non-negative scalars for the regularization parameter.
The default is \code{NULL}, which generates its own \code{lambda} sequence based on
\code{nlambda} and \code{lambda.min.ratio}. For \code{method = "clime"}, the
\code{lambda} sequence is based on \code{nlambda}, \code{lambda.min} and \code{lambda.max}.}

\item{nlambda}{An integer (default = 50) specifying the number of \code{lambda} values
to be generated when \code{lambda = NULL}.}

\item{lambda.min.ratio}{A scalar specifying the fraction of the maximum \code{lambda}
value \eqn{\lambda_{max}} to generate the minimum \code{lambda} \eqn{\lambda_{min}}.
If \code{lambda = NULL}, the program automatically generates a \code{lambda} grid as a
sequence of length \code{nlambda} in log scale, starting from \eqn{\lambda_{min}} to
\eqn{\lambda_{max}}, expect for \code{method = "clime"}. The default value is 0.4 for
\code{method = "tiger"} and 0.01 for others.}

\item{lambda.min}{A scalar specifying the minimum value of program generated
\code{lambda} grid for \code{method = "clime"}. Default is 1e-4 (\eqn{n>p}) or 1e-2
(\eqn{n<p}).}

\item{lambda.max}{A scalar (default = 0.8) specifying the maximum value of program
generated \code{lambda} grid for \code{method = "clime"}.}

\item{gamma}{A scalar specifying the hyperparameter for the chosen \code{method}.
Default values: \enumerate{
\item "elnet": a sequence from 0.1 to 0.9 with increments of 0.1
\item "adapt": 0.5
\item "atan": 0.005
\item "exp": 0.01
\item "scad": 3.7
\item "mcp": 3
}}

\item{target}{A p-by-p symmetric matrix or a scalar (default = 0) serves as the value
for all elements, specifying the target matrix for \code{method = "ridge"} or
\code{method = "elnet"}.}

\item{initial}{A character string (default = "glasso") specifying \itemize{
\item The initial estimate for \code{method} set to \code{"atan"}, \code{"exp"},
\code{"scad"}, or \code{"mcp"}. \itemize{
\item "glasso": use the precision matrix estimate derived from the graphical lasso.
\item "Fan": use the inverse calculation base matrix for \eqn{p < n}; use the precision
matrix estimate derived from the graphical lasso for \eqn{p \geq n}.
}
\item The adaptive weight for \code{method = "adapt"}, calculated as
\eqn{|\tilde{\omega}_{ij}|^{-\gamma}}, where \eqn{\tilde{\Omega} := (\tilde{\omega}_{ij})}.
\itemize{
\item "glasso": use the precision matrix estimate derived from the graphical lasso as
\eqn{\tilde{\Omega}}.
\item "Fan": use the inverse calculation base matrix as \eqn{\tilde{\Omega}} for
\eqn{p < n};  use the precision matrix estimate derived from the graphical lasso as
\eqn{\tilde{\Omega}} for \eqn{p \geq n}.
}
}}

\item{crit}{A string specifying the tuning parameter selection criterion to use.
Available options include: \enumerate{
\item "AIC": Akaike information criterion \insertCite{akaike1973information}{fcstat}.
\item "BIC": Bayesian information criterion \insertCite{schwarz1978estimating}{fcstat}.
\item "EBIC": extended Bayesian information criterion \insertCite{foygel2010extended}{fcstat}.
\item "HBIC": high dimensional Bayesian information criterion \insertCite{wang2013calibrating,fan2017high}{fcstat}.
}}

\item{ebic.tuning}{A scalar (default = 0.5) specifying the tuning parameter to
calculate when \code{crit = "EBIC"}.}

\item{scaling.ratio}{A scalar (default = 0.5) specifying the scaling ratio to adjust
the \code{lambda} grid by multiplying the minimum and maximum with the scaling ratio
for the case where the optimal regularization parameter \code{lambda} is found at the
first index of \code{lambda} grid, make the process run iteratively with the expanded
grid until the optimal parameter is no longer at the first index.}
}
\value{
A list containing the following components: \describe{
\item{hatOmega_opt}{Estimated precision matrix.}
\item{lambda_opt}{Optimal regularization parameter.}
\item{gamma_opt}{Optimal hyperparameter.}
\item{score_opt}{Information criterion score.}
\item{lambda}{Actual lambda grid used in the program.}
\item{gamma}{Actual gamma grid used in the program.}
}
}
\description{
Provide a collection of statistical methods to estimate a precision matrix.
}
\references{
\insertAllCited{}
}
