% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcstat.sel.R
\name{fcstat.sel}
\alias{fcstat.sel}
\title{Selection for precision matrix estimates}
\usage{
fcstat.sel(est.obj, crit = "CV", fold = 5, ebic.tuning = 0.5)
}
\arguments{
\item{est.obj}{The \code{fcstat.est} object output from \code{fcstat.est()}.}

\item{crit}{A string (default = "CV") specifying the parameter selection method to use.
Available options include: \enumerate{
\item "AIC": Akaike information criterion \insertCite{akaike1973information}{fcstat}.
\item "BIC": Bayesian information criterion \insertCite{schwarz1978estimating}{fcstat}.
\item "EBIC": extended Bayesian information criterion \insertCite{foygel2010extended}{fcstat}.
\item "HBIC": high dimensional Bayesian information criterion \insertCite{wang2013calibrating,fan2017high}{fcstat}.
\item "CV": k-fold cross validation with negative log-likelihood loss.
}}

\item{fold}{An integer (default = 5) specifying the number of folds used for \code{crit = "CV"}.}

\item{ebic.tuning}{A scalar (default = 0.5) specifying the tuning parameter to
calculate for \code{crit = "EBIC"}.}
}
\value{
\itemize{
\item For \code{crit = "CV"}, an object with S3 class "fcstat.sel" containing the
following components: \describe{
\item{hatOmega_opt}{The estimated precision matrix.}
\item{lambda_opt}{The optimal regularization parameter.}
\item{gamma_opt}{The optimal hyperparameter.}
\item{loss_opt}{The optimal k-fold loss.}
\item{lambda}{The actual lambda grid used in the program.}
\item{gamma}{The actual gamma grid used in the program.}
\item{loss.mean}{The mean of k-fold loss for each parameter grid value.}
\item{loss.sd}{The standard deviation of k-fold loss for each parameter grid value.}
}
\item For other criteria, an object with S3 class "fcstat.sel" containing the following
components: \describe{
\item{hatOmega_opt}{The estimated precision matrix.}
\item{lambda_opt}{The optimal regularization parameter.}
\item{gamma_opt}{The optimal hyperparameter.}
\item{score_opt}{The optimal information criterion score.}
\item{lambda}{The actual lambda grid used in the program.}
\item{gamma}{The actual gamma grid used in the program.}
\item{score}{The information criterion score for each parameter grid value.}
}
}
}
\description{
Provide a collection of methods to perform parameter selection for \code{fcstat.est} object.
}
\references{
\insertAllCited{}
}
