% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncv_pen.R
\name{ncv_pen}
\alias{ncv_pen}
\title{Non-Convex Regularization Penalty}
\usage{
ncv_pen(penalty, Omega, lambda, gamma)
}
\arguments{
\item{penalty}{A character string specifying the non-convex penalty to use.
Available options include: \enumerate{
\item "atan": Arctangent type penalty \insertCite{wang2016variable}{spice}.
\item "exp": Exponential type penalty \insertCite{wang2018variable}{spice}.
\item "mcp": Minimax concave penalty \insertCite{zou2006adaptive}{spice}.
\item "scad": Smoothly clipped absolute deviation
\insertCite{fan2001variable,fan2009network}{spice}.
}}

\item{Omega}{The precision matrix.}

\item{lambda}{A non-negative numeric value specifying the regularization
parameter.}

\item{gamma}{A numeric value specifying the additional parameter for
the penalty function. The defaults are:
\enumerate{
\item "atan": 0.005
\item "exp": 0.01
\item "mcp": 3
\item "scad": 3.7
}}
}
\value{
A numeric matrix.
}
\description{
Compute a specified non-convex regularization penalty.
}
\references{
\insertAllCited{}
}
