% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcstat.R
\name{fcstat}
\alias{fcstat}
\title{Statistical methods for estimating precision matrix}
\usage{
fcstat(
  X,
  method,
  base = "cov",
  n = NULL,
  lambda = NULL,
  nlambda = 20,
  lambda.min.ratio = 0.01,
  gamma = NA,
  initial = "glasso",
  utilopt = "glassoFast",
  crit = "CV",
  fold = 5,
  ebic.tuning = 0.5,
  cores = 1
)
}
\arguments{
\item{X}{\enumerate{
\item An n-by-p data matrix with sample size n and dimension p.
\item A p-by-p sample covariance/correlation matrix with dimension p.
}}

\item{method}{A character string specifying the statistical method for estimating
precision matrix. Available options include: \enumerate{
\item "glasso": graphical lasso \insertCite{friedman2008sparse}{fcstat}.
\item "ridge": graphical ridge \insertCite{vanwieringen2016ridge}{fcstat}.
\item "elnet": graphical elastic net \insertCite{zou2005regularization}{fcstat}.
\item "clime": constrained L1-minimization for inverse (covariance) matrix estimation
\insertCite{cai2011aconstrained}{fcstat}.
\item "tiger": tuning-insensitive graph estimation and regression
\insertCite{liu2017tiger}{fcstat}, which is only applicable when \code{X} is the n-by-p
data matrix.
\item "adapt": adaptive lasso \insertCite{zou2006adaptive,fan2009network}{fcstat}.
\item "atan": arctangent type penalty \insertCite{wang2016variable}{fcstat}.
\item "exp": exponential type penalty \insertCite{wang2018variable}{fcstat}.
\item "mcp": minimax concave penalty \insertCite{zou2006adaptive}{fcstat}.
\item "scad": smoothly clipped absolute deviation \insertCite{fan2001variable,fan2009network}{fcstat}.
}}

\item{base}{A character string (default = "cov") specifying the calculation base,
either the covariance matrix ("cov") or the correlation matrix ("cor"). This is only
applicable when \code{X} is the n-by-p data matrix.}

\item{n}{An integer (default = NULL) specifying the sample size. This is only required
when the input matrix \code{X} is a p-by-p sample covariance/correlation matrix with
dimension p.}

\item{lambda}{Grid of non-negative scalars for the regularization parameter.
The default is \code{NULL}, which generates its own \code{lambda} sequence based on
\code{nlambda} and \code{lambda.min.ratio}.}

\item{nlambda}{An integer (default = 20) specifying the number of \code{lambda} values
to be generated when \code{lambda = NULL}.}

\item{lambda.min.ratio}{A scalar (default = 0.01) specifying the fraction of
the maximum \code{lambda} value \eqn{\lambda_{max}} to generate the minimum
\code{lambda} \eqn{\lambda_{min}}. If \code{lambda = NULL}, the program automatically
generates a \code{lambda} grid as a sequence of length \code{nlambda} in log scale,
starting from \eqn{\lambda_{min}} to \eqn{\lambda_{max}}.}

\item{gamma}{Grid of scalars specifying the hyperparameter for the chosen \code{method}.
Default values: \enumerate{
\item "elnet": a sequence from 0.1 to 0.9 with increments of 0.1
\item "adapt": 0.5
\item "atan": 0.005
\item "exp": 0.01
\item "scad": 3.7
\item "mcp": 3
}}

\item{initial}{A p-by-p matrix or a p-by-p-by-npara (the number of all combinations of
\code{lambda} and \code{gamma}) array specifying the initial estimate for \code{method}
set to \code{"atan"}, \code{"exp"}, \code{"scad"}, and \code{"mcp"}; or specifying
\eqn{\tilde{\Omega}} of the adaptive weight for \code{method = "adapt"}, calculated as
\eqn{|\tilde{\omega}_{ij}|^{-\gamma}}, where \eqn{\tilde{\Omega} := (\tilde{\omega}_{ij})}.
Some options are also offered when a character string is provided (default "linshrink"),
including:
\itemize{
\item "glasso": use the precision matrix estimate derived from the graphical lasso.
\item "invS": use the inverse calculation base matrix if the matrix is invertible.
\item "linshrink": use the precision matrix estimate derived from Ledoit-Wolf linear
shrinakge estimator of the population covariance matrix
\insertCite{ledoit2004well}{fcstat}.
\item "nlshrink": use the precision matrix estimate derived from Ledoit-Wolf non-linear
shrinakge estimator of the population covariance matrix
\insertCite{ledoit2015spectrum,ledoit2017numerical}{fcstat}.
}}

\item{utilopt}{A character string specifying the utility option to use. The available
options depend on the chosen method: \enumerate{
\item For \code{method = "glasso"}: \itemize{
\item "ADMMsigma": the utility function from \code{\link[ADMMsigma]{ADMMsigma}}.
\item "CVglasso": the utility function from \code{\link[CVglasso]{CVglasso}}.
\item "CovTools": the utility function from \code{\link[CovTools]{PreEst.glasso}}.
\item "glasso": the utility function from \code{\link[glasso]{glasso}}.
\item "GLassoElnetFast": the utility function from
\href{https://github.com/TobiasRuckstuhl/GLassoElnetFast}{gelnet}.
\item "glassoFast": the utility function from \code{\link[glassoFast]{glassoFast}}.
\item "huge": the utility function from \code{\link[huge]{huge.glasso}}.
}
\item For \code{method = "ridge"}: \itemize{
\item "ADMMsigma": the utility function from \code{\link[ADMMsigma]{ADMMsigma}}.
\item "GLassoElnetFast": the utility function from
\href{https://github.com/TobiasRuckstuhl/GLassoElnetFast}{gelnet}.
\item "porridge": the utility function from \code{\link[porridge]{ridgePgen}}.
\item "rags2ridges": the utility function from \code{\link[rags2ridges]{ridgeP}}.
}
\item For \code{method = "elnet"}: \itemize{
\item "ADMMsigma": the utility function from \code{\link[ADMMsigma]{ADMMsigma}}.
\item "GLassoElnetFast": the utility function from
\href{https://github.com/TobiasRuckstuhl/GLassoElnetFast}{gelnet}.
}
\item For \code{method = "clime"}: \itemize{
\item "clime": the utility function from \code{\link[clime]{clime}}.
\item "flare": the utility function from \code{\link[flare]{sugm}}.
}
\item For \code{method = "tiger"}: \itemize{
\item "flare": the utility function from \code{\link[flare]{sugm}}.
\item "huge": the utility function from \code{\link[huge]{huge.tiger}}.
}
\item For \code{method} set to \code{"adapt"}, \code{"atan"}, \code{"exp"},
\code{"scad"}, and \code{"mcp"}: \itemize{
\item "glasso": the utility function from \code{\link[glasso]{glasso}}.
\item "glassoFast": the utility function from \code{\link[glassoFast]{glassoFast}}.
}
}}

\item{crit}{A string (default = "CV") specifying the parameter selection method to use.
Available options include: \enumerate{
\item "AIC": Akaike information criterion \insertCite{akaike1973information}{fcstat}.
\item "BIC": Bayesian information criterion \insertCite{schwarz1978estimating}{fcstat}.
\item "EBIC": extended Bayesian information criterion \insertCite{foygel2010extended}{fcstat}.
\item "HBIC": high dimensional Bayesian information criterion \insertCite{wang2013calibrating,fan2017high}{fcstat}.
\item "CV": k-fold cross validation with negative log-likelihood loss.
}}

\item{fold}{An integer (default = 5) specifying the number of folds used for \code{crit = "CV"}.}

\item{ebic.tuning}{A scalar (default = 0.5) specifying the tuning parameter to
calculate for \code{crit = "EBIC"}.}

\item{cores}{An integer (default = 1) specifying the number of cores to use for
parallel execution.}
}
\value{
\itemize{
\item For \code{crit = "CV"}, an object with S3 class "fcstat" containing the following
components: \describe{
\item{hatOmega_opt}{The estimated precision matrix.}
\item{lambda_opt}{The optimal regularization parameter.}
\item{gamma_opt}{The optimal hyperparameter.}
\item{loss_opt}{The optimal k-fold loss.}
\item{hatOmega}{A list of estimated precision matrices for \code{lambda} grid and
\code{gamma} grid.}
\item{lambda}{The actual lambda grid used in the program.}
\item{gamma}{The actual gamma grid used in the program.}
\item{loss.mean}{The mean of k-fold loss for each parameter grid value.}
\item{loss.sd}{The standard deviation of k-fold loss for each parameter grid value.}
}
\item For other criteria, an object with S3 class "fcstat" containing the following
components: \describe{
\item{hatOmega_opt}{The estimated precision matrix.}
\item{lambda_opt}{The optimal regularization parameter.}
\item{gamma_opt}{The optimal hyperparameter.}
\item{score_opt}{The optimal information criterion score.}
\item{hatOmega}{A list of estimated precision matrices for \code{lambda} grid and
\code{gamma} grid.}
\item{lambda}{The actual lambda grid used in the program.}
\item{gamma}{The actual gamma grid used in the program.}
\item{score}{The information criterion score for each parameter grid value.}
}
}
}
\description{
Provide a collection of statistical methods to estimate a precision matrix.
}
\note{
For the method \code{tiger}, the estimation process solely relies on the raw n-by-p
data \code{X} and does not utilize the argument \code{base}. This argument is not
applicable for \code{tiger} and will have no effect if provided.
}
\references{
\insertAllCited{}
}
