% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ledoit_wolf_est.R
\name{ledoit_wolf_est}
\alias{ledoit_wolf_est}
\title{Ledoit-Wolf shrinkage estimator}
\usage{
ledoit_wolf_est(X, method = "lin", res = "cov")
}
\arguments{
\item{X}{A data matrix.}

\item{method}{A string (default = "lin") specifying the method used in shrinkage,
includes: \enumerate{
\item Linear shrinkage: "lin" (Ledoit and Wolf, 2004)
\item Non-linear shrinkage (Ledoit and Wolf, 2015, 2017): \itemize{
\item "nlminb": non-linear shrinakge using the optimization routine "nlminb".
\item "nloptr": non-linear shrainkage using the optimization routine "nloptr".
}
}
See \code{\link[stats]{nlminb}} and \code{\link[nloptr]{nloptr}} for details.}

\item{res}{A string (default = "cov") specifying the result matrix to be obtained,
either the covariance matrix ("cov") or the correlation matrix ("cor").}
}
\value{
A numeric matrix.
}
\description{
Compute the Ledoit-Wolf shrinkage estimator for the covariance or correlation matrix.
}
\references{
\itemize{
\item Ledoit, Olivier and Wolf, Michael. (2004).
A Well-Conditioned Estimator for Large-Dimensional Covariance Matrices.
\emph{Journal of Multivariate Analysis}, 88(2), 365--411.
\item Ledoit, Olivier and Wolf, Michael. (2015).
Spectrum Estimation: A Unified Framework for Covariance Matrix Estimation and PCA in
Large dimensions.
\emph{Journal of Multivariate Analysis}, 139, 360--384.
\item Ledoit, Olivier and Wolf, Michael. (2017).
Numerical Implementation of the QuEST Function.
\emph{Computational Statistics & Data Analysis}, 115, 199--223.
}
}
