% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{performance}
\alias{performance}
\title{Performance Measures}
\usage{
performance(hatOmega, Omega)
}
\arguments{
\item{hatOmega}{The estimated precision matrix.}

\item{Omega}{The reference precision matrix.}
}
\value{
A list containing the following components: \describe{
\item{Fnorm}{Frobenius (Hilbert-Schmidt) norm between the true and estimated
precision matrices.}
\item{KL}{Kullback-Leibler divergence between the true and estimated
precision matrices.}
\item{Snorm}{Spectral (operator) norm of the difference between the true and
estimated precision matrices.}
\item{precision}{Precision measure, the ratio of true positives to the total
predicted positives.}
\item{recall}{Recall measure, also known as Sensitivity, the ratio of true
positives to the total actual positives.}
\item{specificity}{Specificity measure, the ratio of true negatives to
the total actual negatives.}
\item{F1}{F1 score, the harmonic mean of Precision and Recall.}
\item{MCC}{Matthews correlation coefficient, a measure of the quality of
binary classifications.}
\item{sparsity}{The proportion of zeros among edges in the estimated
precision matrix.}
}
}
\description{
Calculate various measures to evaluate the performance of the estimated
precision matrix.
}
